refactor(mods): seperate mods module into various modules
