+++ app/models/mods/performance.py
@@ -1,117 +1,17 @@
+"""Ranked-mod configuration and validation for PP eligibility checks."""
+
 import hashlib
 import json
-from typing import Any, Literal, NotRequired, TypedDict
+from typing import Any, Literal
 
 from app.config import settings as app_settings
 from app.log import log
 from app.path import CONFIG_DIR, STATIC_DIR
 
+from .definition import API_MODS, APIMod
+
 from pydantic import ConfigDict, Field, create_model
 from pydantic.main import BaseModel
-
-
-class APIMod(TypedDict):
-    acronym: str
-    settings: NotRequired[dict[str, bool | float | str | int]]
-
-
-# https://github.com/ppy/osu-api/wiki#mods
-API_MOD_TO_LEGACY: dict[str, int] = {
-    "NF": 1 << 0,  # No Fail
-    "EZ": 1 << 1,  # Easy
-    "TD": 1 << 2,  # Touch Device
-    "HD": 1 << 3,  # Hidden
-    "HR": 1 << 4,  # Hard Rock
-    "SD": 1 << 5,  # Sudden Death
-    "DT": 1 << 6,  # Double Time
-    "RX": 1 << 7,  # Relax
-    "HT": 1 << 8,  # Half Time
-    "NC": 1 << 9,  # Nightcore
-    "FL": 1 << 10,  # Flashlight
-    "AT": 1 << 11,  # Autoplay
-    "SO": 1 << 12,  # Spun Out
-    "AP": 1 << 13,  # Auto Pilot
-    "PF": 1 << 14,  # Perfect
-    "4K": 1 << 15,  # 4K
-    "5K": 1 << 16,  # 5K
-    "6K": 1 << 17,  # 6K
-    "7K": 1 << 18,  # 7K
-    "8K": 1 << 19,  # 8K
-    "FI": 1 << 20,  # Fade In
-    "RD": 1 << 21,  # Random
-    "CN": 1 << 22,  # Cinema
-    "TP": 1 << 23,  # Target Practice
-    "9K": 1 << 24,  # 9K
-    "CO": 1 << 25,  # Key Co-op
-    "1K": 1 << 26,  # 1K
-    "3K": 1 << 27,  # 3K
-    "2K": 1 << 28,  # 2K
-    "SV2": 1 << 29,  # ScoreV2
-    "MR": 1 << 30,  # Mirror
-}
-LEGACY_MOD_TO_API_MOD = {}
-FREEMOD = 0
-for k, v in API_MOD_TO_LEGACY.items():
-    LEGACY_MOD_TO_API_MOD[v] = APIMod(acronym=k, settings={})
-    FREEMOD |= v
-API_MOD_TO_LEGACY["NC"] |= API_MOD_TO_LEGACY["DT"]
-API_MOD_TO_LEGACY["PF"] |= API_MOD_TO_LEGACY["SD"]
-
-
-# see static/mods.json
-class Settings(TypedDict):
-    Name: str
-    Type: str
-    Label: str
-    Description: str
-
-
-class Mod(TypedDict):
-    Acronym: str
-    Name: str
-    Description: str
-    Type: str
-    Settings: list[Settings]
-    IncompatibleMods: list[str]
-    RequiresConfiguration: bool
-    UserPlayable: bool
-    ValidForMultiplayer: bool
-    ValidForFreestyleAsRequiredMod: bool
-    ValidForMultiplayerAsFreeMod: bool
-    AlwaysValidForSubmission: bool
-
-
-API_MODS: dict[Literal[0, 1, 2, 3], dict[str, Mod]] = {}
-
-
-def init_mods():
-    mods_file = STATIC_DIR / "mods.json"
-    raw_mods = json.loads(mods_file.read_text(encoding="utf-8"))
-    for ruleset in raw_mods:
-        ruleset_mods = {}
-        for mod in ruleset["Mods"]:
-            ruleset_mods[mod["Acronym"]] = mod
-        API_MODS[ruleset["RulesetID"]] = ruleset_mods
-
-
-def int_to_mods(mods: int) -> list[APIMod]:
-    mod_list = []
-    for mod in range(31):
-        if mods & (1 << mod):
-            mod_list.append(LEGACY_MOD_TO_API_MOD[(1 << mod)])
-    if mods & (1 << 14):
-        mod_list.remove(LEGACY_MOD_TO_API_MOD[(1 << 5)])
-    if mods & (1 << 9):
-        mod_list.remove(LEGACY_MOD_TO_API_MOD[(1 << 6)])
-    return mod_list
-
-
-def mods_to_int(mods: list[APIMod]) -> int:
-    sum_ = 0
-    for mod in mods:
-        sum_ |= API_MOD_TO_LEGACY.get(mod["acronym"], 0)
-    return sum_
-
 
 DEFAULT_RANKED_MODS = {
     0: {
@@ -249,10 +149,14 @@ RANKED_MODS: RankedMods = {}
 
 
 class _LegacyModSettings(BaseModel):
+    """Backward-compatible settings view for legacy env flags."""
+
     enable_all_mods_pp: bool = False
 
 
 def _get_mods_file_checksum() -> str:
+    """Return md5 checksum for static mods metadata used by ranked config."""
+
     current_mods_file = STATIC_DIR / "mods.json"
     if not current_mods_file.exists():
         return ""
@@ -260,6 +164,8 @@ def _get_mods_file_checksum() -> str:
 
 
 def generate_ranked_mod_settings(enable_all: bool = False):
+    """Generate config/ranked_mods.json from defaults or all discovered mods."""
+
     ranked_mods_file = CONFIG_DIR / "ranked_mods.json"
     checksum = _get_mods_file_checksum()
     legacy_setting = _LegacyModSettings.model_validate(app_settings.model_dump())
@@ -278,6 +184,8 @@ def generate_ranked_mod_settings(enable_all: bool = False):
 
 
 def init_ranked_mods():
+    """Load ranked mod rules from config and validate against mods checksum."""
+
     ranked_mods_file = CONFIG_DIR / "ranked_mods.json"
     if ranked_mods_file.exists():
         raw_ranked_mods = json.loads(ranked_mods_file.read_text(encoding="utf-8"))
@@ -296,6 +204,8 @@ def init_ranked_mods():
 
 
 def _generate_model(settings: dict[str, dict[str, Any]]) -> type[BaseModel]:
+    """Build a runtime Pydantic model from ranked-mod setting constraints."""
+
     fields = {}
     for setting, validation in settings.items():
         type_ = validation.get("type")
@@ -323,6 +233,8 @@ def _generate_model(settings: dict[str, dict[str, Any]]) -> type[BaseModel]:
 
 
 def check_settings(mod: APIMod, ranked_mods: RulesetRankedMods) -> bool:
+    """Validate one mod settings payload against ranked-mod constraints."""
+
     if (settings := ranked_mods.get(mod["acronym"])) is None:
         return False
     if settings == {}:
@@ -336,6 +248,8 @@ def check_settings(mod: APIMod, ranked_mods: RulesetRankedMods) -> bool:
 
 
 def _mods_can_get_pp(ruleset_id: int, mods: list[APIMod], ranked_mods: RankedMods) -> bool:
+    """Check if a mod list is PP-eligible under a given ranked-mod ruleset."""
+
     for mod in mods:
         if app_settings.enable_rx and mod["acronym"] == "RX" and ruleset_id in {0, 1, 2}:
             continue
@@ -348,51 +262,12 @@ def _mods_can_get_pp(ruleset_id: int, mods: list[APIMod], ranked_mods: RankedMod
 
 
 def mods_can_get_pp_vanilla(ruleset_id: int, mods: list[APIMod]) -> bool:
+    """Validate PP eligibility against built-in default ranked-mod rules."""
+
     return _mods_can_get_pp(ruleset_id, mods, DEFAULT_RANKED_MODS)
 
 
 def mods_can_get_pp(ruleset_id: int, mods: list[APIMod]) -> bool:
+    """Validate PP eligibility against loaded runtime ranked-mod rules."""
+
     return _mods_can_get_pp(ruleset_id, mods, RANKED_MODS)
-
-
-def mod_to_save(mods: list[APIMod]) -> list[str]:
-    s = list({mod["acronym"] for mod in mods})
-    s.sort()
-    return s
-
-
-def get_speed_rate(mods: list[APIMod]):
-    rate = 1.0
-    for mod in mods:
-        if mod["acronym"] in {"DT", "NC", "HT", "DC"}:
-            rate *= mod.get("settings", {}).get("speed_change", 1.0)  # pyright: ignore[reportOperatorIssue]
-    return rate
-
-
-def get_available_mods(ruleset_id: int, required_mods: list[APIMod]) -> list[APIMod]:
-    if ruleset_id not in API_MODS:
-        return []
-
-    ruleset_mods = API_MODS[ruleset_id]
-    required_mod_acronyms = {mod["acronym"] for mod in required_mods}
-
-    incompatible_mods = set()
-    for mod_acronym in required_mod_acronyms:
-        if mod_acronym in ruleset_mods:
-            incompatible_mods.update(ruleset_mods[mod_acronym]["IncompatibleMods"])
-
-    available_mods = []
-    for mod_acronym, mod_data in ruleset_mods.items():
-        if mod_acronym in required_mod_acronyms:
-            continue
-
-        if mod_acronym in incompatible_mods:
-            continue
-
-        if any(required_acronym in mod_data["IncompatibleMods"] for required_acronym in required_mod_acronyms):
-            continue
-
-        if mod_data.get("UserPlayable", False):
-            available_mods.append(mod_acronym)
-
-    return [APIMod(acronym=acronym) for acronym in available_mods]
